Refactor to common emit helper.  Pdb format dependent upon os.
